# fix map!(f,dest::ChainedVector,src::AbstractVector)