Fixes https://github.com/JuliaData/SentinelArrays.jl/issues/115

With this PR the src is iterated instead of the probably longer destination.

Performance is better on this PR:
```
a,b=ChainedVector([rand(Int,5) for _ in 1:20000]),[1:100000...] # many chunks
@btime map!(i->-i,a,b)
# main: 188us
# pr: 153us

a,b=ChainedVector([[1:100000...]]),[1:100000...] # one chunk
@btime map!(i->-i,a,b)
# main: 171us
# pr: 73.8us